Modify: data setup to MainActivity
성능개선
