+++ presentation/src/main/java/kr/hs/dgsw/smartschool/dodamdodam/features/home/HomeFragment.kt
@@ -35,13 +35,13 @@ class HomeFragment : BaseFragment<FragmentHomeBinding, HomeViewModel>() {
     lateinit var mealHomeAdapter: MealHomeAdapter
 
     override fun observerViewModel() {
-        viewModel.dataSetUp()
         viewModel.getAllowSong()
+        viewModel.getMyStudyRoom()
         setLocationRecyclerView()
         setUpTodaySong()
         setMealListViewPager()
+        collectMyStudyRoom()
         collectMealState()
-        collectDataSetUpDate()
         collectSongList()
         initViewEvent()
     }
@@ -52,23 +52,6 @@ class HomeFragment : BaseFragment<FragmentHomeBinding, HomeViewModel>() {
                 HomeViewModel.ON_CLICK_MEAL_MORE -> (activity as? MainActivity)?.moveHomeToMeal()
                 HomeViewModel.ON_CLICK_SONG_MORE -> (activity as? MainActivity)?.moveHomeToSong()
                 HomeViewModel.ON_CLICK_OUT -> findNavController().navigate(R.id.action_main_home_to_outFragment)
-            }
-        }
-    }
-
-    private fun collectDataSetUpDate() {
-        with(viewModel) {
-            lifecycleScope.launchWhenStarted {
-                dataSetUpState.collect { state ->
-                    if (state.result != null) {
-                        getMyStudyRoom()
-                        collectMyStudyRoom()
-                    }
-
-                    if (state.error.isNotBlank()) {
-                        shortToast(state.error)
-                    }
-                }
             }
         }
     }
